fix docker, move urg.test to e2e/
