+++ Dockerfile
@@ -53,6 +53,7 @@ RUN bun run compile:hardhat
 # Expose ports for L1 and L2
 EXPOSE 8545
 EXPOSE 8546
+EXPOSE 8547
 
 # Run devnet
 WORKDIR /app/contracts
