+++ docker-compose.yml
@@ -7,9 +7,11 @@ services:
     ports:
       - "8545:8545" # L1 chain
       - "8546:8546" # L2 chain
+      - "8547:8547" # Urg
     expose:
       - 8545
       - 8546
+      - 8547
     environment:
       ANVIL_IP_ADDR: "0.0.0.0"
     healthcheck:
